+++ packages/location_web/lib/location_web.dart
@@ -91,6 +91,12 @@ class LocationWebPlugin extends LocationPlatform {
   }
 
   @override
+  Future<bool> isBackgroundPermissionGranted() async {
+    // The web platform has no notion of background location permission.
+    return false;
+  }
+
+  @override
   Future<bool> requestService() async {
     return true;
   }